refactor(shared): extract useAuth composable
Both apps had byte-identical Better Auth wrappers. Move to @qzr/shared
as createAppAuthClient(baseURL) factory returning { authClient, useAuth }.
Each app's local useAuth.ts shrinks to factory call + (scoresheet only)
the IS_TAURI constant.

better-auth + vue added as peerDependencies on shared since the
factory's return shape leaks Better Auth's types.

Adds skipLibCheck to shared tsconfig to tolerate better-auth pulling in
node:sqlite / bun:sqlite type refs.
